+++ packages/grpc/src/protobuf/cerbos/effect/v1/effect.ts
@@ -1,7 +1,4 @@
 /* eslint-disable */
-import Long from "long";
-import * as _m0 from "protobufjs/minimal";
-
 export const protobufPackage = "cerbos.effect.v1";
 
 export enum Effect {
@@ -9,9 +6,4 @@ export enum Effect {
   EFFECT_ALLOW = 1,
   EFFECT_DENY = 2,
   EFFECT_NO_MATCH = 3,
-}
-
-if (_m0.util.Long !== Long) {
-  _m0.util.Long = Long as any;
-  _m0.configure();
 }