+++ packages/grpc/src/protobuf/google/api/expr/v1alpha1/syntax.ts
@@ -1,7 +1,7 @@
 /* eslint-disable */
+import { Timestamp } from "../../../../google/protobuf/timestamp";
 import Long from "long";
 import * as _m0 from "protobufjs/minimal";
-import { Timestamp } from "../../../../google/protobuf/timestamp";
 import { NullValue } from "../../../../google/protobuf/struct";
 import { Duration } from "../../../../google/protobuf/duration";