+++ packages/grpc/CHANGELOG.md
@@ -1,6 +1,11 @@
 ## [Unreleased]
 
-No notable changes.
+### Added
+
+- Support for schema validation in `Client#planResources` ([#123](https://github.com/cerbos/cerbos-sdk-javascript/pull/123))
+
+  Requires Cerbos 0.19+.
+  `PlanResourcesResponse#validationErrors` will always return an empty array if the client is connected to an earlier version of Cerbos.
 
 ## [0.5.2] - 2022-06-09